Speech Dictionary dialogs: Split into their on `gui.speechDict` module
As a preparatory step for PR #13294 the code should be moved from gui\settingsDialogs
